// flow-client/src/main/resources/META-INF/resources/frontend/form/BinderState.ts
/*
 * Copyright 2000-2020 Vaadin Ltd.
 *
 * Licensed under the Apache License, Version 2.0 (the "License"); you may not
 * use this file except in compliance with the License. You may obtain a copy of
 * the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
 * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 * License for the specific language governing permissions and limitations under
 * the License.
 */

import {AbstractModel} from "./Models";
import {Validator, ValueError} from "./Validation";

export interface BinderState<T, M extends AbstractModel<T>> {
  readonly model: M;
  readonly name: string;
  value: T;
  readonly defaultValue: T;
  readonly validators: ReadonlyArray<Validator<T>>;
  readonly visited: boolean;
  readonly invalid: boolean;
  readonly errors: ReadonlyArray<ValueError<T>>;

  validate(): Promise<void>;
}